fix: correct mistakes
