+++ .eslintrc.js
@@ -17,6 +17,5 @@ module.exports = {
         sourceType: "module"
     },
     rules: {
-        "linebreak-style": "off",
     }
 };
