+++ .gitignore
@@ -1,7 +1,5 @@
 node_modules
-.DS_Store
 build
 .vscode
 
-Coverage
 coverage
