+++ config/jest/jest.config.js
@@ -10,7 +10,7 @@ module.exports = {
       branches: 100,
       functions: 100,
       lines: 100,
-      statements: -10,
+      statements: 100,
     },
   },
   moduleNameMapper: {
