+++ package.json
@@ -4,10 +4,10 @@
   "description": "react homework",
   "main": "index.js",
   "scripts": {
-    "start": "cross-env NODE_ENV=development && node server.js",
+    "start": "cross-env NODE_ENV=development node server.js",
     "build": "npm run build:clear && webpack --mode production --config config/webpack/webpack.prod.config.js",
     "build:clear": "rm -rf build",
-    "prod": "node server.js",
+    "prod": "npm run build:clear && webpack --mode production --config config/webpack/webpack.prod.config.js && node server.js",
     "lint": "eslint ./src/ && stylelint ./src/**/*.scss",
     "test": "jest --config ./config/jest/jest.config.js --coverage"
   },
@@ -32,7 +32,7 @@
     "react-dom": "^16.12.0"
   },
   "engines": {
-    "node": ">=4.3.2"
+    "node": "13.8.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.4",
@@ -46,7 +46,6 @@
     "css-loader": "^3.4.2",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
-    "eslint-loader": "^3.0.3",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.18.3",
@@ -68,12 +67,10 @@
     "stylelint-config-standard-scss": "^1.1.0",
     "stylelint-scss": "^3.14.2",
     "uglifyjs-webpack-plugin": "^2.2.0",
-    "url-loader": "^3.0.0",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
     "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "^1.7.2"
+    "webpack-merge": "^4.2.2"
   }
 }
