+++ src/App.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Signature from './components/Signature';
+
+const App = () => (
+  <div>
+    <Signature name="Tatsiana Krautsova" />
+  </div>
+);
+
+export default App;
