--- src/components/App/App.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import Signature from '../Signature';
-
-const App = () => (
-  <div>
-    <Signature name="Tatsiana Krautsova" />
-  </div>
-);
-
-export default App;
