--- src/components/App/__test__/app.test.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ShallowRenderer from 'react-test-renderer/shallow';
-import App from '..';
-
-test('App renders correctly', () => {
-  const renderer = new ShallowRenderer();
-  renderer.render(<App />);
-  const result = renderer.getRenderOutput();
-  expect(result).toMatchSnapshot();
-});
