+++ src/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 
 import './main.scss';
 
-import App from './components/App';
+import App from './App';
 
 ReactDOM.render(
   <App />,
